The code is changed so that the Voronoi cells can be cut by hexagon or circle

+++ VoronoiDiagram.h
@@ -54,7 +54,7 @@ public:
 	virtual ~VoronoiDiagram();
 	std::vector<polygon_2d> getVoronoiPolygons(double *XInFrame, double *YInFrame, double *VInFrame, int numPedsInFrame);
 	std::vector<polygon_2d> cutPolygonsWithGeometry(std::vector<polygon_2d> polygon, polygon_2d Geometry, double* xs, double* ys);
-	std::vector<polygon_2d> cutPolygonsWithCircle(std::vector<polygon_2d> polygon, float* xs, float* ys, float radius);
+	std::vector<polygon_2d> cutPolygonsWithCircle(std::vector<polygon_2d> polygon, double* xs, double* ys, double radius);
 };
 
 #endif /* VORONOIDIAGRAM_H_ */
